==> notebooks/02_transform_raw_data_into_ts_data.ipynb <==
# Aggregate Raw Data and Fill Missing Slots

In this notebook we are going to performa the following tasks:

1. Aggregate raw data: function `agg_raw_data_by_hour`
2. Fill missing slots with zeros: function `add_missing_slots`

import pandas as pd

rides = pd.read_parquet('../data/transformed/validated_rides_2022_01.parquet')

rides

rides['pickup_hour'] = rides['pickup_datetime'].dt.floor('H')
rides

agg_rides = rides.groupby(['pickup_hour', 'pickup_location_id']).size().reset_index(name='rides')
agg_rides

# src/data.py
def agg_raw_data_by_hour(rides: pd.DataFrame) -> pd.DataFrame:
    rides['pickup_hour'] = rides['pickup_datetime'].dt.floor('H')
    agg_rides = rides.groupby(['pickup_hour', 'pickup_location_id']).size().reset_index(name='rides')
    return agg_rides

agg_rides = agg_raw_data_by_hour(rides)
agg_rides

from tqdm import tqdm

# src/data.py
def add_missing_slots(agg_rides: pd.DataFrame) -> pd.DataFrame:
    """Funtion to add missing slots on time series data, and fill them with zeroes.

    Args:
        agg_rides (pd.DataFrame): Time Series DataFrame.

    Returns:
        pd.DataFrame: Time Series DataFrame with missing slots filled with zeroes.

    Examples:
        >>> agg_rides = pd.DataFrame({
        ...     'pickup_hour': ['2022-01-01 00:00:00', '2022-01-01 01:00:00', '2022-01-01 03:00:00'],
        ...     'pickup_location_id': [1, 1, 1],
        ...     'rides': [1, 1, 1]
        ... })
        >>> agg_rides
                pickup_hour  pickup_location_id  rides
        0 2022-01-01 00:00:00                   1      1
        1 2022-01-01 01:00:00                   1      1
        2 2022-01-01 03:00:00                   1      1
        >>> add_missing_slots(agg_rides)
                pickup_hour  pickup_location_id  rides
        0 2022-01-01 00:00:00                   1      1
        1 2022-01-01 01:00:00                   1      1
        2 2022-01-01 02:00:00                   1      0
        3 2022-01-01 03:00:00                   1      1
        
    """

    location_ids = agg_rides['pickup_location_id'].unique()
    all_hours = pd.date_range(start=agg_rides['pickup_hour'].min(), end=agg_rides['pickup_hour'].max(), freq='H')

    output = pd.DataFrame()

    for location_id in tqdm(location_ids):

        # keep only rides for this 'location_id'
        agg_rides_i = agg_rides.loc[agg_rides['pickup_location_id'] == location_id, ['pickup_hour', 'rides']]

        # quick way to add missing dates with 0 in a Series
        # taken from https://stackoverflow.com/a/19324591
        agg_rides_i.set_index('pickup_hour', inplace=True)
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)
        agg_rides_i = agg_rides_i.reindex(all_hours, fill_value=0)

        # add back 'location_id' column
        agg_rides_i['pickup_location_id'] = location_id

        output = pd.concat([output, agg_rides_i])

    # move the purchase_day from index to column
    output = output.reset_index().rename(columns={'index': 'pickup_hour'})

    return output

agg_rides_all_slots = add_missing_slots(agg_rides)

agg_rides_all_slots

Now let's pack everything on the function `transform_raw_data_into_ts_data`

# src/data.py
def transform_raw_data_into_ts_data(rides: pd.DataFrame) -> pd.DataFrame:
    # sum rides per location and pickup_hour
    agg_rides = agg_raw_data_by_hour(rides)

    # add rows for (locations, pickup_hours)s with 0 rides
    agg_rides_all_slots = add_missing_slots(agg_rides)

    return agg_rides_all_slots

agg_rides_all_slots = transform_raw_data_into_ts_data(rides)
agg_rides_all_slots

from typing import Optional, List
import plotly.express as px

def plot_rides(rides: pd.DataFrame, locations: Optional[List[int]]=None):
    """Plot time-series data"""

    rides_to_plot = rides[rides.pickup_location_id.isin(locations)] if locations else rides

    fig = px.line(
        rides_to_plot,
        x='pickup_hour',
        y='rides',
        color='pickup_location_id',
        template='none'
    )

    fig.show()

plot_rides(agg_rides_all_slots, locations=[43])

agg_rides_all_slots.to_parquet('../data/transformed/ts_data_2022_01.parquet')
