==> notebooks/03_transform_ts_data_into_features_and_target.ipynb <==
import pandas as pd

ts_data = pd.read_parquet("../data/transformed/ts_data_2022_01.parquet")

ts_data

ts_data_one_location = ts_data.loc[ts_data.pickup_location_id == 43, :].reset_index(drop=True)
ts_data_one_location

def get_cutoff_indices_features_and_target(data: pd.DataFrame, n_features: int, step_size: int) -> list:
    """Function to get the indices for the cutoffs of a Time Series DataFrame.
The Time Series DataFrame should be orderded by time.

    Args:
        data (pd.DataFrame): Time Series DataFrame
        n_features (int): Number of features to use for the prediction
        step_size (int): Step size to use to slide the Time Series DataFrame

    Returns:
        list: list of tuples, which one containing the cuttof indices for the Time Series DataFrame

    Example:
        >>> ts_data = pd.DataFrame({"time": [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]})
        >>> get_cutoff_indices(data=ts_data, n_features=3, step_size=2)
        [(0, 3, 4), (2, 5, 6), (4, 7, 8), (6, 9, 10)]
    """

    stop_position = len(data) - 1

    # Start the first sub-sequence at index position 0
    subseq_first_idx = 0
    subseq_mid_idx = n_features
    subseq_last_idx = n_features + 1
    indices = []

    while subseq_last_idx <= stop_position:
        indices.append((subseq_first_idx, subseq_mid_idx, subseq_last_idx))

        subseq_first_idx += step_size
        subseq_mid_idx += step_size
        subseq_last_idx += step_size

    return indices

print(get_cutoff_indices_features_and_target.__doc__)

n_features=24
step_size=1


indices = get_cutoff_indices_features_and_target(
    ts_data_one_location,
    n_features=n_features,
    step_size=step_size
)
indices[:5]

# Now we use the indices to create the features and target
import numpy as np

n_examples = len(indices)
X = np.ndarray(shape=(n_examples, n_features), dtype=np.float32)
y = np.ndarray(shape=(n_examples), dtype=np.float32)
pickup_hours = []

for i, idx in enumerate(indices):
    X[i, :] = ts_data_one_location.iloc[idx[0]:idx[1]]['rides'].values
    y[i] = ts_data_one_location.iloc[idx[1]:idx[2]]['rides'].values
    pickup_hours.append(ts_data_one_location.iloc[idx[1]]['pickup_hour'])

print(f'{X.shape=}')
print(f'{X=}')
print(f'{pickup_hours[:5]=}')

features_one_location = pd.DataFrame(X, columns=[ f'rides_previous_{i+1}_hour' for i in reversed(range(n_features)) ] )
features_one_location

targets_one_location = pd.DataFrame(y, columns=[f'target_rides_next_hour'])
targets_one_location

from tqdm import tqdm

def transform_ts_data_into_features_and_target(
    ts_data: pd.DataFrame,
    n_features: int,
    step_size: int
) -> pd.DataFrame:
    """
    Slices and transposes data from time-series format into a (features, target)
    format that we can use to train Supervised ML models
    """
    assert set(ts_data.columns) == {'pickup_hour', 'rides', 'pickup_location_id'}

    location_ids = ts_data['pickup_location_id'].unique()
    features = pd.DataFrame()
    targets = pd.DataFrame()
    
    for location_id in tqdm(location_ids):
        
        # keep only ts data for this `location_id`
        ts_data_one_location = ts_data.loc[
            ts_data.pickup_location_id == location_id, 
            ['pickup_hour', 'rides']
        ]

        # pre-compute cutoff indices to split dataframe rows
        indices = get_cutoff_indices(
            data=ts_data_one_location,
            n_features=n_features,
            step_size=step_size
        )

        # slice and transpose data into numpy arrays for features and targets
        n_examples = len(indices)
        x = np.ndarray(shape=(n_examples, n_features), dtype=np.float32)
        y = np.ndarray(shape=(n_examples), dtype=np.float32)
        pickup_hours = []
        for i, idx in enumerate(indices):
            x[i, :] = ts_data_one_location.iloc[idx[0]:idx[1]]['rides'].values
            y[i] = ts_data_one_location.iloc[idx[1]:idx[2]]['rides'].values
            pickup_hours.append(ts_data_one_location.iloc[idx[1]]['pickup_hour'])

        # numpy -> pandas
        features_one_location = pd.DataFrame(
            x,
            columns=[f'rides_previous_{i+1}_hour' for i in reversed(range(n_features))]
        )
        features_one_location['pickup_hour'] = pickup_hours
        features_one_location['pickup_location_id'] = location_id

        # numpy -> pandas
        targets_one_location = pd.DataFrame(y, columns=[f'target_rides_next_hour'])

        # concatenate results
        features = pd.concat([features, features_one_location])
        targets = pd.concat([targets, targets_one_location])

    features.reset_index(inplace=True, drop=True)
    targets.reset_index(inplace=True, drop=True)

    return features, targets['target_rides_next_hour']

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    n_features=24*7*1, # one week of history
    step_size=24,
)

print(f'{features.shape=}')
print(f'{targets.shape=}')
