==> notebooks/04_transform_raw_data_into_features_and_target.ipynb <==
%load_ext autoreload
%autoreload 2

from src.data import load_raw_data

rides = load_raw_data(year=2022)
rides

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data

Let's build a dataset (ABT) with all the hours included. This is the larger dataset we can build with and also the most representative one.

from src.data import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*28*1, # one month in hours
    step_size=1, # 1 hour in advance
)

print(f'{features.shape=}')
print(f'{targets.shape=}')

tabular_data = features
tabular_data['target_rides_next_hour'] = targets

from src.paths import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / 'tabular_data_step_size_1h.parquet')

tabular_data

Now let's build another dataset less representative, with step_size=23. We are make 23h step hour forward, but we have every hour represented in the dataset.

tabular_data.pickup_hour.dt.hour.value_counts().sort_index()

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*28*1, # one month in hours
    step_size=23, # 1 hour in advance
)

print(f'{features.shape=}')
print(f'{targets.shape=}')

tabular_data = features
tabular_data['target_rides_next_hour'] = targets

from src.paths import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / 'tabular_data_step_size_23h.parquet')

tabular_data.pickup_hour.dt.hour.value_counts().sort_index()
