==> notebooks/11_backfill_feature_store.ipynb <==
import os
from dotenv import load_dotenv
from src.paths import PARENT_DIR

load_dotenv(PARENT_DIR / '.env')

HOPSWORK_PROJECT_NAME = 'taxi_demand_ytjqdg'
HOPWSWORKS_API_KEY = os.environ['HOPSWORKS_API_KEY']

import pandas as pd
from datetime import datetime
from src.data import load_raw_data

from_year = 2022
to_year = datetime.now().year
print(f"Downloading raw data from {from_year} to {to_year}")

rides = pd.DataFrame()
for year in range(from_year, to_year+1):
    # download data for the whole year
    rides_one_year = load_raw_data(year)
    # append rows
    rides = pd.concat([rides, rides_one_year])

rides

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)

ts_data

import hopsworks

project = hopsworks.login(
    project=HOPSWORK_PROJECT_NAME,
    api_key_value=HOPWSWORKS_API_KEY
)

feature_store = project.get_feature_store()

FEATURE_GROUP_NAME = 'time_series_hourly_feature_group'
FEATURE_GROUP_VERSION = 1

feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description="Time-series data at hourly frequency",
    primary_key = ['pickup_location_id', 'pickup_hour'],
    event_time='pickup_hour',
)

feature_group.insert(ts_data, write_options={"wait_for_job": False})

Testando a leitura dos dados que subiram pra feature store

# Create feature view (if it doesn't exist yet)
# This feature view only uses on feature group, so the query is trivial
import src.config as config

try:
    # create feature view if it doesn't exist yet
    feature_store.create_feature_view(
        name=config.FEATURE_VIEW_NAME,
        version=config.FEATURE_VIEW_VERSION,
        query=feature_group.select_all()
    )
except:
    print('Feature view already existed. Skip creation.')


# get feature view
feature_view = feature_store.get_feature_view(
    name=config.FEATURE_VIEW_NAME,
    version=config.FEATURE_VIEW_VERSION
)

ts_data_, _ = feature_view.training_data(
    description='Time-series hourly taxi rides'
)

ts_data_.sort_values(by=['pickup_location_id', 'pickup_hour'])

ts_data_, _ = feature_view.training_data(
    description='Time-series hourly taxi rides'
)
ts_data_.sort_values(by=['pickup_location_id', 'pickup_hour'])
